Raise precision on size percentage calculation on summary
Mainly to fix a bug where equal original and final sizes can
sometimes in slight percentage mismatch such as:
| original: 717565 (100.000000%)
| cp: 717565 (100.000000%)
This also should help to make the percentage a bit more accurate

+++ internal/compressor/compress.go
@@ -525,7 +525,7 @@ func writeSizeLine(summaryBuilder *strings.Builder, result *CompressionResult, i
 		return
 	}
 
-	percentage := float32(result.FinalSize) * 100 / float32(result.OriginalSize)
+	percentage := float64(result.FinalSize) * 100 / float64(result.OriginalSize)
 	sizeLine := fmt.Sprintf("%d (%f%%)", result.FinalSize, percentage)
 	if isBest {
 		sizeLine = color.GreenString(sizeLine)
